+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [[tbd]] (2023-08-22)
+
+### Added
+- LT-4955: Add new publishing strategy 'PropertiesWithMessageTypeTopicPublishStrategy'
+
 ## 11.9.3 (2023-08-10)
 
 ### Added